Moved MMS.java, MMSTest.java, and README.md to root; removed MMS-Java-Split folder
