+++ README.md
@@ -1,6 +1,6 @@
 # MMS: Memory Management System (Java Simulation)
 
-This project simulates a simple Memory Management System (MMS) using linked lists in Java.
+This project simulates a simple Memory Management System (MMS) using linked lists in Java. It was created as part of an academic assignment at York University for the course ITEC 2620 (Summer 2025).
 
 ---
 
@@ -19,10 +19,24 @@ The MMS manages memory allocation and deallocation requests from processes using
 - **Garbage Collection**: Automatically compacts memory when no block is large enough due to fragmentation.
 - **Sorted Allocation List**: Maintains `allocList` in increasing order of address for simpler garbage collection.
 
----
+## 🛠️ Usage
 
-## 🛠️ How to Compile and Run
-
-### 1. Compile:
+### Compile
 ```bash
 javac MMS.java
+```
+
+### Run
+```bash
+java MMS
+```
+
+### Commands
+```text
+alloc <name> <size>
+free <name>
+printalloc
+printfree
+exit
+```
+
